requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.2.4
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
